+++ packages/yoastseo/src/languageProcessing/languages/es/config/internal/passiveVoiceAuxiliaries.js
