+++ packages/yoastseo/src/languageProcessing/languages/es/config/internal/passiveVoiceParticiples.js
